Separated settings in its own module
