+++ app/react/Settings/components/DocumentTypesList.js
