+++ app/react/Settings/components/specs/CollectionSettings.spec.js
