+++ app/react/Settings/components/specs/SettingsNavigation.spec.js
