+++ app/react/Users/reducer.js
@@ -3,7 +3,6 @@ import createReducer from 'app/BasicReducer';
 import {modelReducer, formReducer} from 'react-redux-form';
 
 export default combineReducers({
-  section: createReducer('users/section', 'account'),
-  form: modelReducer('users.form', {username: '', password: ''}),
-  formState: formReducer('users.form')
+  form: modelReducer('login.form', {username: '', password: ''}),
+  formState: formReducer('login.form')
 });
