+++ app/react/reducer.js
@@ -1,6 +1,5 @@
 import {combineReducers} from 'redux';
 import {reducer as reduxForm} from 'redux-form';
-import createReducer from 'app/BasicReducer';
 
 import templates from 'app/Templates/reducers/templatesReducer';
 import template from 'app/Templates/reducers/reducer';
@@ -16,7 +15,8 @@ import library from 'app/Library/reducers/reducer';
 import modals from 'app/Modals/reducers/modalsReducer';
 import uploads from 'app/Uploads/reducers/reducer';
 import user from 'app/Auth/reducer';
-import users from 'app/Users/reducers/reducer';
+import settings from 'app/Settings/reducer';
+import login from 'app/Users/reducer';
 
 import {modelReducer, formReducer} from 'react-redux-form';
 
@@ -36,7 +36,7 @@ export default combineReducers({
   modals,
   uploads,
   user,
-  users,
-  settings: createReducer('settings', {}),
+  login,
+  settings,
   search: modelReducer('search', {sort: 'creationDate', order: 'desc', searchTerm: '', filters: {}})
 });
